+++ core/src/main/java/com/contentful/vault/Constants.java
@@ -5,7 +5,9 @@ public final class Constants {
     throw new AssertionError();
   }
 
-  public static final String SUFFIX_SPACE = "$$SpaceHelper";
+  public static final String SUFFIX_FIELDS = "$Fields";
 
   public static final String SUFFIX_MODEL = "$$ModelHelper";
+
+  public static final String SUFFIX_SPACE = "$$SpaceHelper";
 }